requirements.txt
anyio==3.4.0
ariadne==0.14.0
autopep8==1.6.0
cffi==1.15.0
click==8.0.3
Flask==2.0.2
graphql-core==3.1.7
greenlet==1.1.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
psycopg2==2.9.2
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
python-dotenv==0.19.2
sniffio==1.2.0
SQLAlchemy==1.4.29
starlette==0.17.1
toml==0.10.2
typing_extensions==4.0.1
Werkzeug==2.0.2
